+++ src/Core/Window.hpp
@@ -18,8 +18,8 @@ class Game;
 class Window
 {
 public:
-  constexpr inline static uint16_t kUnitsPerScreenX = 100;
-  constexpr inline static uint16_t kUnitsPerScreenY = 100;
+  constexpr inline static int kUnitsPerScreenX = 100;
+  constexpr inline static int kUnitsPerScreenY = 100;
 
   Window(Game &game) noexcept;
   void Exit() noexcept;
@@ -28,11 +28,11 @@ public:
   [[nodiscard]] constexpr inline auto GetWidth() noexcept -> int { return width_; }
   [[nodiscard]] constexpr inline auto GetHeight() noexcept -> int { return height_; }
   void SetResolution(const int width, const int height) noexcept;
-  [[nodiscard]] constexpr inline auto GetRenderWidth() noexcept -> uint16_t { return render_width_; }
-  [[nodiscard]] constexpr inline auto GetRenderHeight() noexcept -> uint16_t { return render_height_; }
-  void SetRenderResolution(const uint16_t render_width, const uint16_t render_height) noexcept;
-  [[nodiscard]] constexpr inline auto GetPixelsPerUnitX() noexcept -> uint16_t { return pixels_per_unit_x_; }
-  [[nodiscard]] constexpr inline auto GetPixelsPerUnitY() noexcept -> uint16_t { return pixels_per_unit_y_; }
+  [[nodiscard]] constexpr inline auto GetRenderWidth() noexcept -> int { return render_width_; }
+  [[nodiscard]] constexpr inline auto GetRenderHeight() noexcept -> int { return render_height_; }
+  void SetRenderResolution(const int render_width, const int render_height) noexcept;
+  [[nodiscard]] constexpr inline auto GetPixelsPerUnitX() noexcept -> int { return pixels_per_unit_x_; }
+  [[nodiscard]] constexpr inline auto GetPixelsPerUnitY() noexcept -> int { return pixels_per_unit_y_; }
   
   [[nodiscard]] constexpr inline auto GetSDLWindow() noexcept -> SDL_Window* { return sdl_window_; }
 
@@ -52,8 +52,8 @@ public:
   /// Use int instead of uint16_t as elsewhere because SDL requires pointers to int to fetch screen resolution
   int width_;
   int height_;
-  uint16_t render_width_ = 0;
-  uint16_t render_height_ = 0;
+  int render_width_ = 0;
+  int render_height_ = 0;
   std::string title_ = "Game";
 
   DefFloatType pixels_per_unit_x_;
